feat: add task module routes and reorganize imports
Add task module route setup to main.go to enable task-related API 
endpoints. Reorganize imports for better clarity and consistency by 
grouping related packages and removing unused ones. This change 
prepares the application for handling task domain features alongside 
existing category functionality.
